Add a developer GUC "enable_datanode_row_triggers" to allow ROW TRIGGERS to be executed on the datanodes.
This must be used with caution. Postgres-XL does not officially support
triggers yet. One of the reasons for not supporting triggers is that a trigger
function executed on a datanode may not have access to all the required data
since the data may not reside on the datanode. But if users are confident that
the triggers can be safely executed on the datanode, they may turn this GUC on.
Still since the feature is not well tested, we don't recommend users to use
this without thorough testing and knowing consequences.