fix set_subquery_pathlist()
Thanks to upper-planner pathification (3fc6e2d7f), subquery_planner
now produces a list of paths instead of a single plan. So tweak the
code to process all the paths.

XXX All the paths share the distribution, apparently. So maybe this
does not make much sense and we should pick one of the paths?