Ensure variable values are quoted when necessary while sending down SET comamnds to the remote nodes
Earlier we'd special cased a few GUCs such as those using memory or time units
or transaction isolation levels. But clearly that wasn't enough as we noticed
with "application_name" recently. So fix this problem in a more comprehensive
manner.

Added a few more test cases to cover these scenarios.