fix SharedQueuesInit to register the tranche properly
The code was mixing code for named and regular tranches (identified
by just ID, in this case LWTRANCHE_SHARED_QUEUES). So switch to
the LWLockRegisterTranche() approach.

There were a two more bugs though:

1) The tranche struct was local in the block, which does not work as
it's referenced in the global array without creating a copy. So just
make it static and global.

2) The locks were not properly initialized. Add a LWLockInitialize
call to the loop.