fix various minor bugs in initdb, to make it work
1) setup_nodeself() was starting it's own postgres process, which
then conflicted with the already running one. Instead just pass
the 'cmdfd' as for the other setup_* methods.

2) make_template0() included VACUUM command with a single '\n'
at the end of the line, resulting in this error:

FATAL:  VACUUM cannot be executed from a function or multi-command string

Apparently, there need to be '\n\n' as in the  non-XL code.